+++ crates/imap/src/op/capability.rs
@@ -11,7 +11,7 @@ use common::network::SessionStream;
 use imap_proto::{
     Command, StatusResponse,
     protocol::{
-        ImapResponse,
+        ImapResponse, quoted_string,
         capability::{Capability, Response},
     },
     receiver::Request,
@@ -70,6 +70,35 @@ impl<T: SessionStream> Session<T> {
                     .as_bytes()
                     .to_vec(),
                 ),
+        )
+        .await
+    }
+
+    pub async fn handle_jmap_access(&mut self, request: Request<Command>) -> trc::Result<()> {
+        // Validate access
+        self.assert_has_permission(Permission::ImapCapability)?;
+
+        let op_start = Instant::now();
+        trc::event!(
+            Imap(trc::ImapEvent::Capabilities),
+            SpanId = self.session_id,
+            Elapsed = op_start.elapsed()
+        );
+
+        let mut response = b"* JMAPACCESS ".to_vec();
+        quoted_string(
+            &mut response,
+            &format!(
+                "{}/.well-known/jmap",
+                self.server.core.network.http.url_https
+            ),
+        );
+        response.extend_from_slice(b"\r\n");
+
+        self.write_bytes(
+            StatusResponse::completed(Command::GetJmapAccess)
+                .with_tag(request.tag)
+                .serialize(response),
         )
         .await
     }